(BSR)[BO] chore: remove 'WIP_ENABLE_BO_COLLECTIVE_OFFER_DETAILS_V2' feature flag
