--- api/src/pcapi/routes/backoffice/templates/collective_offer/details_v2.html
@@ -1,142 +1,0 @@
-{% extends "layouts/details.html" %}
-{% from "components/turbo/lazy_modal.html" import build_lazy_modal with context %}
-{% import "components/badges.html" as badges with context %}
-{% import "components/clipboard.html" as clipboard %}
-{% import "components/links.html" as links with context %}
-{% from "components/connect_as.html" import build_connect_as %}
-{% block title %}{{ collective_offer.name }}{% endblock %}
-{% block header_extra %}<div>{{ collective_offer.validation | format_offer_validation_status(with_badge=True) }}</div>{% endblock %}
-{% block action_buttons %}
-  {% if is_collective_offer_price_editable and has_permission("ADVANCED_PRO_SUPPORT") %}
-    {{ action_bar_button("pencil-square", "Ajuster le prix", modal_id="#update-collective-offer-price") }}
-  {% endif %}
-  {% if has_permission("PRO_FRAUD_ACTIONS") %}
-    {{ action_bar_button("hand-thumbs-up", "Valider", modal_id="#validate-collective-offer-modal-" + collective_offer.id|string) }}
-    {{ action_bar_button("hand-thumbs-down", "Rejeter", modal_id="#reject-collective-offer-modal-" + collective_offer.id|string) }}
-    {% if is_feature_active("MOVE_OFFER_TEST") %}
-      {{ action_bar_button("arrow-left-right", "Déplacer l'offre", modal_id="#move-collective-offer-modal-" + collective_offer.id|string) }}
-    {% endif %}
-  {% endif %}
-{% endblock action_buttons %}
-{% block extra_title_bar %}
-  {% call content_navbar("offer-details") %}
-    {{ content_navbar_element('info', 'Détails') }}
-  {% endcall %}
-{% endblock extra_title_bar %}
-{% block details_container %}
-  {% call details_content_wrapper("offer-details") %}
-    <!-- INFO -->
-    {% call nav_section("info", "Détails de l'offre") %}
-      {% call description_detail_horizontal("Formats") %}
-        {{ collective_offer.formats | format_collective_offer_formats }}
-      {% endcall %}
-      {% if collective_offer.collectiveStock %}
-        {% call description_detail_horizontal("Prix") %}
-          {{ collective_offer.collectiveStock.price | format_amount }}
-        {% endcall %}
-      {% endif %}
-      {% if collective_offer.collectiveStock %}
-        {% call description_detail_horizontal("Places") %}
-          {{ collective_offer.collectiveStock.numberOfTickets }}
-        {% endcall %}
-      {% endif %}
-      {% if collective_offer.institution %}
-        {% call description_detail_horizontal("Établissement") %}
-          {{ collective_offer.institution.name }}
-        {% endcall %}
-      {% endif %}
-      {% if collective_offer.teacher %}
-        {% call description_detail_horizontal("Enseignant") %}
-          {{ collective_offer.teacher.full_name }}
-        {% endcall %}
-      {% endif %}
-      {% call description_detail_horizontal("Date de création") %}
-        {{ collective_offer.dateCreated | format_date }}
-      {% endcall %}
-      {% call description_detail_horizontal("Date de l'évènement") %}
-        {% if collective_offer.collectiveStock %}
-          {% set start_date = collective_offer.start | format_date %}
-          {% set end_date = collective_offer.end | format_date %}
-          {{ start_date }}
-          {% if end_date and end_date != start_date %}→ {{ end_date }}{% endif %}
-        {% endif %}
-      {% endcall %}
-    {% endcall %}
-  {% endcall %}
-{% endblock details_container %}
-{% block side_column %}
-  {% call build_connect_as(connect_as) %}
-    <button type="button"
-            class="btn btn-outline-primary-subtle-bg"
-            href="{{ connect_as.href }}"
-            data-bs-toggle="tooltip"
-            data-bs-placement="top"
-            data-bs-title="Accéder à l'offre sur PC Pro"
-            data-submit-form="{{ connect_as.formName }}">
-      Accéder PC Pro
-      <i class="bi bi-briefcase"></i>
-    </button>
-  {% endcall %}
-  <div>
-    {% call description_detail_vertical("CollectiveOffer ID") %}
-      {{ collective_offer.id }}
-      {{ clipboard.copy_to_clipboard(collective_offer.id, "Copier l'ID de l'offre") }}
-    {% endcall %}
-    {% call description_detail_vertical("Statut") %}
-      {{ collective_offer.status | format_offer_status }}
-    {% endcall %}
-    {% call description_detail_vertical("Statut PC Pro") %}
-      {{ collective_offer.displayedStatus | format_collective_offer_displayed_status }}
-    {% endcall %}
-    <hr />
-    {% call description_detail_vertical("Entité juridique") %}
-      <div class="vstack gap-1">
-        <div>{{ links.build_offerer_name_to_details_link(collective_offer.venue.managingOfferer) }}</div>
-        {% call badges.badges_container() %}
-          {{ badges.build_offer_offerer_fraud_badges(collective_offer.venue.managingOfferer) }}
-        {% endcall %}
-      </div>
-    {% endcall %}
-    {% call description_detail_vertical("Partenaire culturel") %}
-      <div class="vstack gap-1">
-        <div>{{ links.build_venue_name_to_details_link(collective_offer.venue) }}</div>
-        {% call badges.badges_container() %}
-          {{ badges.build_offer_venue_fraud_badges(collective_offer.venue) }}
-        {% endcall %}
-      </div>
-    {% endcall %}
-    {% if collective_offer.template %}
-      {% call description_detail_vertical("Offre vitrine liée") %}
-        {{ links.build_collective_offer_template_details_link(collective_offer.template, text_attr="name") }}
-      {% endcall %}
-    {% endif %}
-    {% if collective_offer.rejectionReason or collective_offer.lastValidationDate or collective_offer.lastValidationAuthor %}<hr />{% endif %}
-    {% if collective_offer.rejectionReason %}
-      {% call description_detail_vertical("Raison de rejet") %}
-        {{ collective_offer.rejectionReason | format_collective_offer_rejection_reason }}
-      {% endcall %}
-    {% endif %}
-    {% if collective_offer.lastValidationDate %}
-      {% call description_detail_vertical("Date de dernière validation") %}
-        {{ collective_offer.lastValidationDate | format_date_time }}
-      {% endcall %}
-    {% endif %}
-    {% if collective_offer.lastValidationAuthor %}
-      {% call description_detail_vertical("Utilisateur de la dernière validation") %}
-        {{ collective_offer.lastValidationAuthor.full_name }}
-      {% endcall %}
-    {% endif %}
-  </div>
-{% endblock side_column %}
-{% block extra_main_content %}
-  {% if is_collective_offer_price_editable and has_permission("ADVANCED_PRO_SUPPORT") %}
-    {{ build_lazy_modal(url_for("backoffice_web.collective_offer.get_collective_offer_price_form", collective_offer_id=collective_offer.id) , "update-collective-offer-price") }}
-  {% endif %}
-  {% if has_permission("PRO_FRAUD_ACTIONS") %}
-    {{ build_lazy_modal(url_for('backoffice_web.collective_offer.get_validate_collective_offer_form', collective_offer_id=collective_offer.id) , "validate-collective-offer-modal-" + collective_offer.id|string) }}
-    {{ build_lazy_modal(url_for('backoffice_web.collective_offer.get_reject_collective_offer_form', collective_offer_id=collective_offer.id) , "reject-collective-offer-modal-" + collective_offer.id|string) }}
-    {% if is_feature_active("MOVE_OFFER_TEST") %}
-      {{ build_lazy_modal(url_for('backoffice_web.collective_offer.get_move_collective_offer_form', collective_offer_id=collective_offer.id) , "move-collective-offer-modal-" + collective_offer.id|string) }}
-    {% endif %}
-  {% endif %}
-{% endblock extra_main_content %}
